+++ tests/files/src/0815_isset_on_prop.php
@@ -1,0 +1,13 @@
+<?php
+class X {
+    public $f;
+    public function test(array $values) {
+        $this->f = [];
+        foreach ($values as $k => $v) {
+            if (!isset($this->f[$v]) ||
+                $this->f[$v] < $k) { // Should not emit PhanTypeInvalidDimOffset.
+                $this->f[$v] = $k;
+            }
+        }
+    }
+}