+++ NEWS.md
@@ -9,7 +9,11 @@ New features(Analysis):
 + Emit `PhanTypeInvalidDimOffset` for accessing any dimension on an empty string or an empty array. (#3385)
 + Warn about invalid string literal offsets such as `'str'[3]`, `'str'[-4]`, etc. (#3385)
 + Infer that arrays are non-empty and support array access from `isset($x[$offset])` (#3463)
-+ Make `array_key_exists` imply array is non-empty. (#3465)
++ Make `array_key_exists` imply argument is a `non-empty-array` (or an `object`). (#3465, #3469)
++ Make `isset($x[$offset])` imply argument is a `non-empty-array`, `object`, or `string`
+  Make `isset($x['literal string'])` imply argument is a `non-empty-array` or `object`, and not a `string`.
++ Make `isset($x->prop)` imply `$x` is an `object`.
++ Make `isset($this->prop[$x])` imply `$this->prop` is not the empty array shape. (#3467)
 
 Nov 03 2019, Phan 2.4.1
 -----------------------